Rename dist_autograd_context and dist_autograd_container.
The paths distributed/autograd/context/dist_autograd_context.h and
distributed/autograd/context/dist_autograd_container.h were repetitive.

Therefore renaming these to distributed/autograd/context/context.h and
distributed/autograd/context/container.h

Differential Revision: [D18467624](https://our.internmc.facebook.com/intern/diff/D18467624/)

[ghstack-poisoned]
