+++ tools/build_variables.py
@@ -51,8 +51,8 @@ libtorch_sources = [
     "torch/csrc/autograd/saved_variable.cpp",
     "torch/csrc/autograd/variable.cpp",
     "torch/csrc/distributed/autograd/utils.cpp",
-    "torch/csrc/distributed/autograd/context/dist_autograd_container.cpp",
-    "torch/csrc/distributed/autograd/context/dist_autograd_context.cpp",
+    "torch/csrc/distributed/autograd/context/container.cpp",
+    "torch/csrc/distributed/autograd/context/context.cpp",
     "torch/csrc/distributed/autograd/engine/dist_engine.cpp",
     "torch/csrc/distributed/autograd/functions/recvrpc_backward.cpp",
     "torch/csrc/distributed/autograd/functions/sendrpc_backward.cpp",
