+++ torch/csrc/distributed/autograd/context/context.cpp
@@ -1,6 +1,6 @@
 #include <functional>
 
-#include <torch/csrc/distributed/autograd/context/dist_autograd_context.h>
+#include <torch/csrc/distributed/autograd/context/context.h>
 #include <c10/util/Exception.h>
 
 namespace torch {
